editors - move open with to editor (fix #100057)
