+++ src/vs/workbench/services/editor/common/editorOpenWith.ts
@@ -3,20 +3,24 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { basename, extname, isEqual } from 'vs/base/common/resources';
-import { URI } from 'vs/base/common/uri';
 import * as nls from 'vs/nls';
-import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IEditorOptions, ITextEditorOptions } from 'vs/platform/editor/common/editor';
-import { IQuickInputService, IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
-import { IEditorInput, IEditorPane } from 'vs/workbench/common/editor';
-import { FileEditorInput } from 'vs/workbench/contrib/files/common/editors/fileEditorInput';
-import { DEFAULT_EDITOR_ID } from 'vs/workbench/contrib/files/common/files';
-import { CustomEditorAssociation, CustomEditorsAssociations, customEditorsAssociationsSettingId } from 'vs/workbench/services/editor/common/editorAssociationsSetting';
+import { IJSONSchema } from 'vs/base/common/jsonSchema';
+import { IConfigurationNode, IConfigurationRegistry, Extensions } from 'vs/platform/configuration/common/configurationRegistry';
+import { workbenchConfigurationNodeBase } from 'vs/workbench/common/configuration';
+import { Registry } from 'vs/platform/registry/common/platform';
+import { ICustomEditorInfo, IEditorService, IOpenEditorOverrideHandler, IOpenEditorOverrideEntry } from 'vs/workbench/services/editor/common/editorService';
+import { IEditorInput, IEditorPane, IEditorInputFactoryRegistry, Extensions as EditorExtensions } from 'vs/workbench/common/editor';
+import { ITextEditorOptions, IEditorOptions } from 'vs/platform/editor/common/editor';
 import { IEditorGroup, OpenEditorContext } from 'vs/workbench/services/editor/common/editorGroupsService';
-import { IEditorService, IOpenEditorOverrideEntry, IOpenEditorOverrideHandler } from 'vs/workbench/services/editor/common/editorService';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { IQuickInputService, IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
+import { URI } from 'vs/base/common/uri';
+import { extname, basename, isEqual } from 'vs/base/common/resources';
 
-const builtinProviderDisplayName = nls.localize('builtinProviderDisplayName', "Built-in");
+/**
+ * Id of the default editor for open with.
+ */
+export const DEFAULT_EDITOR_ID = 'default';
 
 /**
  * Try to open an resource with a given editor.
@@ -111,10 +115,12 @@ export async function openEditorWith(
 	return pickedItem?.handler.open(input, { ...options, override: pickedItem.id }, group, OpenEditorContext.NEW_EDITOR)?.override;
 }
 
+const builtinProviderDisplayName = nls.localize('builtinProviderDisplayName', "Built-in");
+
 export const defaultEditorOverrideEntry = Object.freeze({
 	id: DEFAULT_EDITOR_ID,
 	label: nls.localize('promptOpenWith.defaultEditor.displayName', "Text Editor"),
-	detail: builtinProviderDisplayName,
+	detail: builtinProviderDisplayName
 });
 
 /**
@@ -124,8 +130,9 @@ export function getAllAvailableEditors(
 	resource: URI,
 	options: IEditorOptions | ITextEditorOptions | undefined,
 	group: IEditorGroup,
-	editorService: IEditorService,
+	editorService: IEditorService
 ): Array<[IOpenEditorOverrideHandler, IOpenEditorOverrideEntry]> {
+	const fileEditorInputFactory = Registry.as<IEditorInputFactoryRegistry>(EditorExtensions.EditorInputFactories).getFileEditorInputFactory();
 	const overrides = editorService.getEditorOverrides(resource, options, group);
 	if (!overrides.some(([_, entry]) => entry.id === DEFAULT_EDITOR_ID)) {
 		overrides.unshift([
@@ -142,9 +149,71 @@ export function getAllAvailableEditors(
 			},
 			{
 				...defaultEditorOverrideEntry,
-				active: editorService.activeEditor instanceof FileEditorInput && isEqual(editorService.activeEditor.resource, resource),
+				active: fileEditorInputFactory.isFileEditorInput(editorService.activeEditor) && isEqual(editorService.activeEditor.resource, resource),
 			}]);
 	}
+
 	return overrides;
 }
 
+export const customEditorsAssociationsSettingId = 'workbench.editorAssociations';
+
+export const viewTypeSchamaAddition: IJSONSchema = {
+	type: 'string',
+	enum: []
+};
+
+export type CustomEditorAssociation = {
+	readonly viewType: string;
+	readonly filenamePattern?: string;
+};
+
+export type CustomEditorsAssociations = readonly CustomEditorAssociation[];
+
+export const editorAssociationsConfigurationNode: IConfigurationNode = {
+	...workbenchConfigurationNodeBase,
+	properties: {
+		[customEditorsAssociationsSettingId]: {
+			type: 'array',
+			markdownDescription: nls.localize('editor.editorAssociations', "Configure which editor to use for specific file types."),
+			items: {
+				type: 'object',
+				defaultSnippets: [{
+					body: {
+						'viewType': '$1',
+						'filenamePattern': '$2'
+					}
+				}],
+				properties: {
+					'viewType': {
+						anyOf: [
+							{
+								type: 'string',
+								description: nls.localize('editor.editorAssociations.viewType', "The unique id of the editor to use."),
+							},
+							viewTypeSchamaAddition
+						]
+					},
+					'filenamePattern': {
+						type: 'string',
+						description: nls.localize('editor.editorAssociations.filenamePattern', "Glob pattern specifying which files the editor should be used for."),
+					}
+				}
+			}
+		}
+	}
+};
+
+export const DEFAULT_CUSTOM_EDITOR: ICustomEditorInfo = {
+	id: 'default',
+	displayName: nls.localize('promptOpenWith.defaultEditor.displayName', "Text Editor"),
+	providerDisplayName: builtinProviderDisplayName
+};
+
+export function updateViewTypeSchema(enumValues: string[], enumDescriptions: string[]): void {
+	viewTypeSchamaAddition.enum = enumValues;
+	viewTypeSchamaAddition.enumDescriptions = enumDescriptions;
+
+	Registry.as<IConfigurationRegistry>(Extensions.Configuration)
+		.notifyConfigurationSchemaUpdated(editorAssociationsConfigurationNode);
+}
